database-service rearranged
