+++ skedge-backend/coursesTesting.json
@@ -25,5 +25,33 @@
             "classEndDate": "19/08/2014",
             "career": "Undergraduate",
             "enrollmentCapacity": "72"
-        }    
+        },
+        {
+            "session": "13W",
+            "subject": "BALBALBLAL2",
+            "catalog": "201",
+            "section": "50",
+            "componentCode": "LEC",
+            "courseTitle": "INTRODUCTORY BIOLOGY",
+            "locationCode": "LOY",
+            "instructionModeCode": "P",
+            "instructionModeDescription": "In Person",
+            "meetingPatternNumber": "1",
+            "roomCode": "CC320",
+            "buildingCode": "CC",
+            "room": "320",
+            "classStartTime": "18.30.00",
+            "classEndTime": "21.00.00",
+            "modays": "N",
+            "tuesdays": "Y",
+            "wednesdays": "N",
+            "thursdays": null,
+            "fridays": "N",
+            "saturdays": "N",
+            "sundays": "N",
+            "classStartDate": "07/05/2014",
+            "classEndDate": "19/08/2014",
+            "career": "Undergraduate",
+            "enrollmentCapacity": "72"
+        }       
 ]
