+++ skedge-backend/database-service.js
@@ -3,34 +3,34 @@
 // Everything related to connecting to MongoDB and manipulating the database
 //
 //
-const mongoose = require('mongoose');
-const Schema = mongoose.Schema;
-const ObjectId = Schema.ObjectId;
-const assert = require('assert');
-let courseModel = require('./courseSchem');
-fs = require('fs');
+// const mongoose = require('mongoose');
+// const Schema = mongoose.Schema;
+// const ObjectId = Schema.ObjectId;
+// const assert = require('assert');
+// let courseModel = require('./courseSchem');
+// fs = require('fs');
 
 
-const server = 'skedge-user:8sDBuOw3zMD4ZpQp@skedge-cantaloop-kueik.mongodb.net';
-const database = 'skedge-app';
+// const server = 'skedge-user:8sDBuOw3zMD4ZpQp@skedge-cantaloop-kueik.mongodb.net';
+// const database = 'skedge-app';
 
 
-class Database {
-    constructor() {
-      this._connect()
-    }
-  _connect() {
-       mongoose.connect("mongodb+srv://skedge-user:8sDBuOw3zMD4ZpQp@skedge-cantaloop-kueik.mongodb.net/skedge-app")
-         .then(() => {
-           console.log('Database connection successful')
-         })
-         .catch(err => {
-           console.error('Database connection error')
-         })
-    }
-  }
+// class Database {
+//     constructor() {
+//       this._connect()
+//     }
+//   _connect() {
+//        mongoose.connect("mongodb+srv://skedge-user:8sDBuOw3zMD4ZpQp@skedge-cantaloop-kueik.mongodb.net/skedge-app")
+//          .then(() => {
+//            console.log('Database connection successful')
+//          })
+//          .catch(err => {
+//            console.error('Database connection error')
+//          })
+//     }
+//   }
 
-  module.exports = new Database();
+//   module.exports = new Database();
 
     // const db = new Database();
 
@@ -75,7 +75,7 @@ class Database {
 //         console.log(doc)
 //     })
 //     .catch( err => {
-//         console.log("errorr")
+//         console.log("error")
 //     })
 
 
@@ -91,36 +91,15 @@ var listConverted = JSON.parse(listCourse);
 
 mongoose.connection.collection('courses').insertMany(listConverted, function( err, result){
     if(err){
-        console.log("error ca marche pas");
+        console.log("Error, fail");
     }else{
         console.log("yay reussi");
     }
 })
 
-// for (i in listConverted){
-//     for( attri in i){
-//         if (typeof(listConverted.columns[i,attri]) != 'undefined'){
-//         var blabla = new courseModel(listConverted.columns[i,attri]);
-//         blabla.save(function (err){
-//             if (err) return console.log("ca pas marcher!!");
-//         })
-//     }
-// }
-// };
 
 
 
-const insertDocuments = function(db, callback){
-	const collection = db.collection('courses');
-	collection.insertMany([{a:1}, {a:2}], 
-	function (err, result){
-		assert.equal(err,null);
-		assert.equal(2, result.result.n);
-		assert.equal(2, result.ops.length);
-		console.log("Inserted 2 courses successfully");
-		callback(result);
-	});
-}
